requirements.txt
torch>=2.1
torchvision>=0.16
transformers>=4.40
numpy>=1.24
pillow>=10
tqdm>=4.66
pyyaml>=6
modal>=0.64
